Begin implementing Transaction, Account Type, User and Category controllers.
Genericize Repository and Service interfaces with out of the box GUI functions

+++ Models/AccountHistory.cs
@@ -6,14 +6,16 @@ namespace AssetManager.Models
     {
         public int HistoryId { get; set; }
         public int AccountId { get; set; }
-        public float CurrentBalance { get; set; }
+        public decimal CurrentBalance { get; set; }
         public DateTime ChangeDate { get; set; }
+
+        //System.Int32 HistoryId, System.Int32 AccountId, System.Decimal CurrentBalance, System.DateTime ChangeDate
 
         [JsonConstructor]
         public AccountHistory(
-            int historyId, 
-            int accountId, 
-            float currentBalance, 
+            int historyId,
+            int accountId,
+            decimal currentBalance,
             DateTime changeDate
         )
         {
